Move text-utils.ts to /common
