+++ extensions/ql-vscode/src/common/text-utils.ts
