+++ extensions/ql-vscode/src/view/variant-analysis/RawResultsTable.tsx
@@ -9,7 +9,7 @@ import {
 } from "../../common/bqrs-cli-types";
 import { tryGetRemoteLocation } from "../../common/bqrs-utils";
 import TextButton from "../common/TextButton";
-import { convertNonPrintableChars } from "../../pure/text-utils";
+import { convertNonPrintableChars } from "../../common/text-utils";
 import { sendTelemetry, useTelemetryOnChange } from "../common/telemetry";
 
 const numOfResultsInContractedMode = 5;
